Move realtime handler to its own file
